+++ backend/src/main/java/mouda/backend/auth/business/KakaoAuthService.java
@@ -3,9 +3,9 @@ package mouda.backend.auth.business;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
-import mouda.backend.auth.implement.JwtProvider;
 import mouda.backend.auth.implement.KakaoOauthManager;
 import mouda.backend.auth.implement.LoginManager;
+import mouda.backend.auth.implement.jwt.AccessTokenProvider;
 import mouda.backend.auth.presentation.request.OauthRequest;
 import mouda.backend.auth.presentation.response.LoginResponse;
 import mouda.backend.member.domain.LoginDetail;
@@ -18,7 +18,7 @@ import mouda.backend.member.implement.MemberWriter;
 @RequiredArgsConstructor
 public class KakaoAuthService implements AuthService {
 
-	private final JwtProvider jwtProvider;
+	private final AccessTokenProvider accessTokenProvider;
 	private final KakaoOauthManager oauthManager;
 	private final LoginManager loginManager;
 	private final MemberFinder memberFinder;
@@ -31,18 +31,18 @@ public class KakaoAuthService implements AuthService {
 	}
 
 	public Member findMember(String token) {
-		long memberId = jwtProvider.extractMemberId(token);
+		long memberId = accessTokenProvider.extractMemberId(token);
 		return memberFinder.find(memberId);
 	}
 
 	public void checkAuthentication(String token) {
-		jwtProvider.validateExpiration(token);
+		accessTokenProvider.validateExpiration(token);
 	}
 
 	public LoginResponse basicLogin() {
 		Member member = new Member("nickname",
 			new LoginDetail(OauthType.KAKAO, 1L));
 		memberWriter.append(member);
-		return new LoginResponse(jwtProvider.createToken(member));
+		return new LoginResponse(accessTokenProvider.provide(member));
 	}
 }
