+++ backend/src/main/java/mouda/backend/auth/implement/LoginManager.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 import org.springframework.stereotype.Component;
 
 import lombok.RequiredArgsConstructor;
+import mouda.backend.auth.implement.jwt.AccessTokenProvider;
 import mouda.backend.member.domain.LoginDetail;
 import mouda.backend.member.domain.Member;
 import mouda.backend.member.domain.OauthType;
@@ -16,14 +17,14 @@ import mouda.backend.member.infrastructure.MemberRepository;
 public class LoginManager {
 
 	private final MemberRepository memberRepository;
-	private final JwtProvider jwtProvider;
+	private final AccessTokenProvider accessTokenProvider;
 	private final MemberWriter memberWriter;
 
 	public String processKakaoLogin(long kakaoId) {
 		Optional<Member> member = memberRepository.findByLoginDetail_SocialLoginId(kakaoId);
 
 		if (member.isPresent()) {
-			return jwtProvider.createToken(member.get());
+			return accessTokenProvider.provide(member.get());
 		}
 
 		Member newMember = Member.builder()
@@ -32,6 +33,6 @@ public class LoginManager {
 			.build();
 		memberWriter.append(newMember);
 
-		return jwtProvider.createToken(newMember);
+		return accessTokenProvider.provide(newMember);
 	}
 }
